Message passing works in userspace!
The problem with making the processes run in ring3, and not being able to
communicate with each other, was because of how the messages were passed: only
the pointers to them were stored, so after a context switch they were pointing
to 'invalid' memory. The solution was to copy the entire message to the
process' mailbox, not just the pointer

By the way, the TODO was right. Who'd have thought? :)